% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight_ab.R
\name{calc_weight_ab}
\alias{calc_weight_ab}
\title{Calculate Fish Weights from Survey As and Bs}
\usage{
calc_weight_ab(data = NULL, ab_source = NULL)
}
\arguments{
\item{data}{A dataframe containing columns for species code, sex, survey season, and length in the format that they are used in the survey database (SPEC,FSEX,SEASON,FLEN). To run for As and Bs, by quarter, it also needs to have a numeric "MONTH" column.}

\item{ab_source}{Source of the ab values: "survey" (Noble and Clark 2017) or "other" which will default to quartly As and Bs as used in Catch.}
}
\value{
"data" input file with additional columns for calculated weight (weight_ab) and weight units
}
\description{
Calculate Fish Weights from Survey As and Bs
}
