% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_in_folder.R
\name{check_files_in_folder}
\alias{check_files_in_folder}
\title{Check Files in Folder}
\usage{
check_files_in_folder(folder_path, file_names)
}
\arguments{
\item{folder_path}{A filepath to the folder where the files should be}

\item{file_names}{The files that should be in the folder}
}
\value{
TRUE (no files missing) or FALSE (missing files)
}
\description{
Check Files in Folder
}
