% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_convert1.R
\name{survey_convert}
\alias{survey_convert}
\title{Apply Groundfish Conversion Factors from the DFO Spring and Summer Surveys}
\usage{
survey_convert(
  summary = "total",
  species = 10,
  season_select = "summer",
  stratum = c(470:478, 480:485, 490:495),
  year_start = 1970,
  year_end = 2024,
  type = 1,
  equivalent = "cartier",
  ab = "survey",
  localdir = NULL,
  login = NULL,
  pw = NULL,
  dsn = NULL
)
}
\arguments{
\item{summary}{The type of summary needed. The options are abundance and biomass per length group ("lf"), set level ("set_level"), strata level ("strata level"), or totals for the selected area ("total"). Default is "total"}

\item{species}{The species codes you want to include. They must require length-based abundance conversion. You can include one or multiple concatenated. Default is cod 10.}

\item{season_select}{The season of the survey. Default is summer.}

\item{stratum}{The selected stratum. Default is 4X.}

\item{year_start}{The first year required. Default is 1970.}

\item{year_end}{The final year required. Default is 2024.}

\item{type}{The type of sets to include. Default is 1, valid primary sets.}

\item{equivalent}{The vessel equivalent. You can select any vessel, but, for example: "needler","teleost","cartier". Default is "cartier"}

\item{ab}{The source of the ab values that are applied. See the function calc_weight_ab. Options include "survey" and "4Xcod" currently. Default is "survey"}

\item{localdir}{A file path for a local directory. Temporarily, you need the file result_conversion.RData in that directory if you want to do spring conversions.}

\item{login}{Your username for the appropriate Oracle schema.}

\item{pw}{Your password for the appropriate Oracle schema.}

\item{dsn}{The name of the schema.}
}
\value{
Dataframe of species, length, from vessel, to vessel (vessel equivalent) and conversion factor.
}
\description{
Apply Groundfish Conversion Factors from the DFO Spring and Summer Surveys
}
