% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_weight_ab.R
\name{calc_weight_ab}
\alias{calc_weight_ab}
\title{Calculate Fish Weights from Survey As and Bs}
\usage{
calc_weight_ab(data = NULL, ab_file = spoctackle::survey_ab)
}
\arguments{
\item{data}{A dataframe containing columns for species code, sex, survey season, and length}

\item{ab_file}{A dataframe containing the survey a and b values}
}
\value{
"data" input file with additional columns for calculated weight (weight_ab) and weight units
}
\description{
Calculate Fish Weights from Survey As and Bs
}
